Separate harness code to reuse in TCK tests
